Ticcmd flags, demo markers, counts and FRACBITS get their real unsigned types
Second batch of signed constants that were only signed because C's int
was. Picked from the per-constant flip experiment (see the previous
commit); each one below had its type chosen from how it is used, and the
ones whose neighbours are genuinely signed were left alone.

u8 (they are bytes on the wire / in the ticcmd):
  BT_ATTACK BT_USE BT_CHANGE BT_SPECIAL BT_SPECIALMASK BT_WEAPONMASK
  BTS_PAUSE BTS_SAVEGAME BTS_SAVEMASK  -- cmd.buttons is a u8, so every
    site widened it to i32 and truncated back; now `cmd.buttons |= BT_USE`
    and `cmd.buttons & BT_SPECIAL != 0`.
  DEMOMARKER SAVEGAME_EOF DOOM_191_VERSION HU_BROADCAST -- compared with
    and written as bytes. vanilla_version_code and demo_version_description
    follow (a demo header byte cannot be negative).
usize (counts / indices):
  NUMPSPRITES NUMAMMO AM_NUMMARKPOINTS MAXLIGHTSCALE SAVESTRINGSIZE
  BODYQUESIZE
  and the fields that only ever indexed with them: markpointnum,
  bodyqueslot, save_char_index. decrease_ammo takes the ammo index as
  usize (its 7 callers cast the AmmoType once, as before).
u32 (shift amounts):
  FRACBITS SLOPEBITS DBITS LIGHTSCALESHIFT MAPBLOCKSHIFT MAPBTOFRAC
  and RMainState::detailshift (0 or 1, only ever a shift amount).

Left signed on purpose, with the reason:
  FINEANGLES ST_HEIGHT SBARHEIGHT STARTREDPALS ST_NUM*FACES HU_FONTSIZE
  TEXTSPEED BACKUPTICS -- used in arithmetic with signed values that can
    go negative (screen coordinates, tic counters, char - HU_FONTSTART);
  MAX_POSITION NUM_QUITMESSAGES -- the unsigned cast is on the *result* of
    a signed expression, not on the constant;
  MAX_JOY_BUTTONS MAX_MOUSE_BUTTONS -- compared with config values that
    may be -1 ("unbound");
  KEY_* -- events carry the key as i32 (Event::data1);
  DOOMGENERIC_RES{X,Y} -- passed through DoomPlatform::init(i32, i32),
    which core_s3 implements;
  FF_FRAMEMASK -- state frame numbers are i32;
  AM_MSGHEADER, SLOPERANGE -- values of i32 fields / return values.

Verified: cargo test --release (57 pass, goldens unchanged), clippy and
rustfmt clean, workspace builds.

+++ rust-doomgeneric/engine/src/d_ticcmd.rs
@@ -1,14 +1,14 @@
 pub const BTS_SAVESHIFT: u32 = 2;
-pub const BTS_SAVEMASK: i32 = 28;
-pub const BTS_SAVEGAME: i32 = 2;
-pub const BTS_PAUSE: i32 = 1;
+pub const BTS_SAVEMASK: u8 = 28;
+pub const BTS_SAVEGAME: u8 = 2;
+pub const BTS_PAUSE: u8 = 1;
 pub const BT_WEAPONSHIFT: u32 = 3;
-pub const BT_WEAPONMASK: i32 = 56;
-pub const BT_CHANGE: i32 = 4;
-pub const BT_SPECIALMASK: i32 = 3;
-pub const BT_SPECIAL: i32 = 128;
-pub const BT_USE: i32 = 2;
-pub const BT_ATTACK: i32 = 1;
+pub const BT_WEAPONMASK: u8 = 56;
+pub const BT_CHANGE: u8 = 4;
+pub const BT_SPECIALMASK: u8 = 3;
+pub const BT_SPECIAL: u8 = 128;
+pub const BT_USE: u8 = 2;
+pub const BT_ATTACK: u8 = 1;
 
 #[derive(Copy, Clone)]
 pub struct TicCmd {
